@@ -3,6 +3,7 @@
   "description": "Check and format markdown or MDX using Prettier as a remark plugin",
   "version": "1.3.0",
   "main": "dist/index.js",
+  "type": "module",
   "files": [
     "dist"
   ],
@@ -29,12 +30,13 @@
   ],
   "scripts": {
     "prepack": "tsc --project tsconfig.build.json",
-    "test": "jest"
+    "test": "NODE_OPTIONS=--experimental-vm-modules jest"
   },
   "dependencies": {
-    "mdast-util-to-markdown": "^0.6.0",
+    "@types/mdast": "^3.0.0",
+    "mdast-util-to-markdown": "^1.0.0",
     "prettier-linter-helpers": "^1.0.0",
-    "vfile-location": "^3.0.0"
+    "vfile-location": "^4.0.0"
   },
   "peerDependencies": {
     "prettier": "^2"
@@ -43,19 +45,19 @@
     "@types/jest": "^27.0.0",
     "@types/prettier-linter-helpers": "^1.0.0",
     "@typescript-eslint/eslint-plugin": "^5.0.0",
-    "eslint": "^7.0.0",
+    "eslint": "^8.0.0",
     "eslint-config-remcohaszing": "^3.0.0",
-    "eslint-plugin-jest": "^25.0.0",
+    "eslint-plugin-jest": "^26.0.0",
     "eslint-plugin-jest-formatting": "^3.0.0",
-    "jest": "^27.0.0",
+    "jest": "^28.0.0-alpha.7",
     "memfs": "^3.0.0",
     "prettier": "^2.0.0",
-    "remark": "^13.0.0",
-    "remark-cli": "^9.0.0",
+    "remark": "^14.0.0",
+    "remark-cli": "^10.0.0",
     "ts-jest": "^27.0.0",
     "ts-node": "^10.0.0",
     "typescript": "^4.0.0",
-    "unified": "^9.0.0",
-    "vfile": "^4.0.0"
+    "unified": "^10.0.0",
+    "vfile": "^5.0.0"
   }
 }
